# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Database System Design Project",
        "contact": {
            "name": "Maintainer Polaris",
            "email": "polaris@fduhole.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "basePath": "/api",
    "paths": {
        "/api/commodities": {
            "put": {
                "description": "更新商品",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "商品"
                ],
                "summary": "更新商品",
                "parameters": [
                    {
                        "description": "json",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Commodity"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "description": "添加商品",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "商品"
                ],
                "summary": "添加商品",
                "parameters": [
                    {
                        "description": "json",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/commodity.CreateCommodityRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "description": "删除商品",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "商品"
                ],
                "summary": "删除商品",
                "parameters": [
                    {
                        "type": "string",
                        "description": "commodity id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/api/commodity/item": {
            "put": {
                "description": "更新商品",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Item"
                ],
                "summary": "更新商品",
                "parameters": [
                    {
                        "description": "json",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/item.UpdateItemModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "description": "添加商品",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Item"
                ],
                "summary": "添加商品",
                "parameters": [
                    {
                        "description": "json",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/item.CreateItemModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "description": "删除商品",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Item"
                ],
                "summary": "删除商品",
                "parameters": [
                    {
                        "type": "string",
                        "description": "commodity id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/favorites": {
            "post": {
                "description": "AddFavorite",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Favorite"
                ],
                "summary": "AddFavorite",
                "parameters": [
                    {
                        "description": "json",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/favorite.AddFavoriteModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/api/platform": {
            "put": {
                "description": "Update platform",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Platform"
                ],
                "summary": "Update platform",
                "parameters": [
                    {
                        "description": "json",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/platform.UpdatePlatformRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "description": "Add platform",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Platform"
                ],
                "summary": "Add platform",
                "parameters": [
                    {
                        "description": "json",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/platform.CreatePlatformRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "description": "Delete platform",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Platform"
                ],
                "summary": "Delete platform",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "platform id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/api/price/history": {
            "put": {
                "description": "Update priceChange",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PriceChange"
                ],
                "summary": "Update priceChange",
                "parameters": [
                    {
                        "description": "json",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/priceChange.UpdatePriceChangeModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "description": "Get priceChange by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PriceChange"
                ],
                "summary": "Get priceChange by ID",
                "parameters": [
                    {
                        "description": "json",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/priceChange.GetPriceChangeModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PriceChange"
                        }
                    }
                }
            }
        },
        "/api/price/limit": {
            "post": {
                "description": "AddPriceLimit",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Favorite"
                ],
                "summary": "AddPriceLimit",
                "parameters": [
                    {
                        "description": "json",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/favorite.PriceLimitModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/search": {
            "post": {
                "description": "搜索商品",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Item"
                ],
                "summary": "搜索商品",
                "parameters": [
                    {
                        "description": "Range: name, category; Search: content to search",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/item.SearchQuery"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CommodityItem"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "commodity.CreateCommodityRequest": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "default_name": {
                    "type": "string"
                },
                "produce_address": {
                    "type": "string"
                },
                "produce_at": {
                    "type": "string"
                }
            }
        },
...
        "favorite.AddFavoriteModel": {
            "type": "object",
            "required": [
                "item_id"
            ],
            "properties": {
                "item_id": {
                    "type": "integer"
                }
            }
        },
        "favorite.PriceLimitModel": {
            "type": "object",
            "properties": {
                "item_id": {
                    "type": "integer"
                },
                "price_limit": {
                    "type": "integer"
                }
            }
        },
        "item.CreateItemModel": {
            "type": "object",
            "required": [
                "commodity_id",
                "item_name",
                "platform_id",
                "price"
            ],
            "properties": {
                "commodity_id": {
                    "type": "integer"
                },
                "item_name": {
                    "type": "string"
                },
                "platform_id": {
                    "type": "integer"
                },
                "price": {
                    "type": "number"
                }
            }
        },
        "item.SearchQuery": {
            "type": "object",
            "properties": {
                "accurate": {
                    "type": "boolean"
                },
                "range": {
                    "type": "string"
                },
                "search": {
                    "type": "string"
                }
            }
        },
        "item.UpdateItemModel": {
            "type": "object",
            "required": [
                "commodity_item_id",
                "item_name",
                "price"
            ],
            "properties": {
                "commodity_item_id": {
                    "type": "integer"
                },
                "item_name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                }
            }
        },
        "models.Commodity": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "default_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "produce_address": {
                    "type": "string"
                },
                "produce_at": {
                    "$ref": "#/definitions/models.MyTime"
                }
            }
        },
        "models.CommodityItem": {
            "type": "object",
            "properties": {
                "commodity": {
                    "$ref": "#/definitions/models.Commodity"
                },
                "commodity_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "item_name": {
                    "type": "string"
                },
                "platform": {
                    "$ref": "#/definitions/models.Platform"
                },
                "platform_id": {
                    "type": "integer"
                },
                "price": {
                    "type": "number"
                },
                "seller": {
                    "$ref": "#/definitions/models.Seller"
                },
                "seller_id": {
                    "type": "integer"
                },
                "update_at": {
                    "$ref": "#/definitions/models.MyTime"
                }
            }
        },
...
        "models.MyTime": {
            "type": "object",
            "properties": {
                "time.Time": {
                    "type": "string"
                }
            }
        },
        "models.Platform": {
            "type": "object",
            "properties": {
                "country": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "models.PriceChange": {
            "type": "object",
            "properties": {
                "commodityItem": {
                    "$ref": "#/definitions/models.CommodityItem"
                },
                "commodity_item_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "new_price": {
                    "type": "number"
                },
                "update_at": {
                    "$ref": "#/definitions/models.MyTime"
                }
            }
        },
        "models.Seller": {
            "type": "object",
            "properties": {
                "address": {
                    "description": "Age      int",
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "platform.CreatePlatformRequest": {
            "type": "object",
            "properties": {
                "country": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "platform.UpdatePlatformRequest": {
            "type": "object",
            "properties": {
                "country": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "priceChange.GetPriceChangeModel": {
            "type": "object",
            "required": [
                "commodity_item_id",
                "time_end",
                "time_start"
            ],
            "properties": {
                "commodity_item_id": {
                    "type": "integer"
                },
                "time_end": {
                    "type": "string"
                },
                "time_start": {
                    "type": "string"
                }
            }
        },
        "priceChange.UpdatePriceChangeModel": {
            "type": "object",
            "properties": {
                "commodity_item_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "new_price": {
                    "type": "number"
                }
            }
        },
...
    }
}